Move packages / artifacts to maven-usable namespaces
