+++ awsagentprovider/src/main/java/software/amazon/opentelemetry/javaagent/providers/AwsOpenTelemetryFactory.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package com.softwareaws.xray.opentelemetry.providers;
+package software.amazon.opentelemetry.javaagent.providers;
 
 import io.opentelemetry.api.OpenTelemetry;
 import io.opentelemetry.sdk.OpenTelemetrySdk;
