+++ awsagentprovider/src/main/resources/META-INF/services/io.opentelemetry.javaagent.shaded.io.opentelemetry.spi.OpenTelemetryFactory
@@ -13,4 +13,4 @@
 # permissions and limitations under the License.
 #
 
-com.softwareaws.xray.opentelemetry.providers.AwsOpenTelemetryFactory
+software.amazon.opentelemetry.javaagent.providers.AwsOpenTelemetryFactory
