+++ awsagentprovider/src/main/resources/META-INF/services/io.opentelemetry.javaagent.spi.config.PropertySource
@@ -1,1 +1,1 @@
-com.softwareaws.xray.opentelemetry.providers.AwsAgentProperties
+software.amazon.opentelemetry.javaagent.providers.AwsAgentProperties
